Chore: --20220407 give standard code format and test case
